@@ -13,7 +13,7 @@ dependencies:
     sdk: flutter
 
   google_fonts: ^2.0.0
-  font_awesome_flutter: ^9.0.0
+  font_awesome_flutter: ^10.1.0
   url_launcher: ^6.0.3
   dough: ^1.0.1
   fluttertoast: ^8.0.6
@@ -28,7 +28,7 @@ dependencies:
   firebase_dynamic_links: ^4.1.0
   image_picker: ^0.8.0+3
   dotted_border: ^2.0.0
-  share_plus: ^3.1.0
+  share_plus: ^4.0.4
   package_info_plus: ^1.0.0
   qr_flutter: ^4.0.0
   dio: ^4.0.0
