@@ -7,7 +7,7 @@ class SocialLists {
     SocialModel('WhatsApp',
         icon: FontAwesomeIcons.whatsapp, colour: Colors.teal.shade800),
     const SocialModel('Phone',
-        icon: FontAwesomeIcons.phoneAlt, colour: Color(0xFF38A45B)),
+        icon: FontAwesomeIcons.phoneFlip, colour: Color(0xFF38A45B)),
     SocialModel('Telegram',
         icon: FontAwesomeIcons.telegram, colour: Colors.blue.shade800),
     const SocialModel('Twitter',
@@ -17,7 +17,7 @@ class SocialLists {
     const SocialModel('Tiktok',
         icon: FontAwesomeIcons.tiktok, colour: Color(0xFFEE1D52)),
     const SocialModel('Sms',
-        icon: FontAwesomeIcons.sms, colour: Color(0xFFDC722B)),
+        icon: FontAwesomeIcons.commentSms, colour: Color(0xFFDC722B)),
     const SocialModel('Email',
         icon: FontAwesomeIcons.at, colour: Color(0xFFBB001B)),
     const SocialModel('YouTube',
@@ -47,9 +47,9 @@ class SocialLists {
     const SocialModel('Amazon',
         icon: FontAwesomeIcons.amazon, colour: Color(0xFFFF9900)),
     const SocialModel('Snapchat',
-        icon: FontAwesomeIcons.snapchatGhost, colour: Colors.black),
+        icon: FontAwesomeIcons.snapchat, colour: Colors.black),
     const SocialModel('Medium',
-        icon: FontAwesomeIcons.mediumM, colour: Color(0xFF000000)),
+        icon: FontAwesomeIcons.medium, colour: Color(0xFF000000)),
     const SocialModel('Pinterest',
         icon: FontAwesomeIcons.pinterestP, colour: Color(0xFFE60023)),
     const SocialModel('WeChat',
@@ -73,15 +73,15 @@ class SocialLists {
     const SocialModel('Behance',
         icon: FontAwesomeIcons.behance, colour: Color(0xFF1666F7)),
     const SocialModel('Shopee',
-        icon: FontAwesomeIcons.shoppingBag, colour: Color(0xFFE64A2E)),
+        icon: FontAwesomeIcons.bagShopping, colour: Color(0xFFE64A2E)),
     SocialModel('User',
-        icon: FontAwesomeIcons.userAlt, colour: Colors.lightGreen.shade700),
+        icon: FontAwesomeIcons.userLarge, colour: Colors.lightGreen.shade700),
     SocialModel('Users',
         icon: FontAwesomeIcons.users, colour: Colors.green.shade700),
     const SocialModel('Blog',
         icon: FontAwesomeIcons.blog, colour: Color(0xFF087E8B)),
     const SocialModel('Blog2',
-        icon: FontAwesomeIcons.penAlt, colour: Color(0xFF087E8B)),
+        icon: FontAwesomeIcons.penClip, colour: Color(0xFF087E8B)),
     const SocialModel('GeneralWebsite',
         icon: FontAwesomeIcons.globe, colour: Colors.blueGrey),
     const SocialModel('GeneralWebsite2',
@@ -89,7 +89,7 @@ class SocialLists {
     const SocialModel('GeneralWebsite3',
         icon: FontAwesomeIcons.link, colour: Colors.blueGrey),
     const SocialModel('GeneralWebsite4',
-        icon: FontAwesomeIcons.poll, colour: Colors.blueGrey),
+        icon: FontAwesomeIcons.squarePollVertical, colour: Colors.blueGrey),
   ];
 
   static SocialModel getSocial(String? exactName) {
