@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_markdown/flutter_markdown.dart';
 
 class HelpDialogs {
   static Widget editModehelpDialog(BuildContext context) {
@@ -10,54 +11,12 @@ class HelpDialogs {
       content: Column(
         mainAxisSize: MainAxisSize.min,
         children: const [
+          UnorderedListItem('**Tap the card** to edit.'),
+          UnorderedListItem('**Swipe right** to delete card.'),
           UnorderedListItem(
-            TextSpan(
-              children: [
-                TextSpan(
-                    text: 'Tap the card',
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                TextSpan(text: ' to edit.\n'),
-              ],
-            ),
-          ),
+              'To reoder/rearrange the card, turn on the **Reoder toggle button**, then hold and drag a card to desired position.'),
           UnorderedListItem(
-            TextSpan(
-              children: [
-                TextSpan(
-                    text: 'Swipe right',
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                TextSpan(text: ' to delete card.\n'),
-              ],
-            ),
-          ),
-          UnorderedListItem(
-            TextSpan(
-              children: [
-                TextSpan(text: 'To reoder/rearrange the card, turn on the'),
-                TextSpan(
-                    text: ' Reoder toggle button',
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                TextSpan(
-                    text: ', then hold and drag a card to desired position.\n'),
-              ],
-            ),
-          ),
-          UnorderedListItem(
-            TextSpan(
-              children: [
-                TextSpan(text: 'Once you\'re done. Switch to '),
-                TextSpan(
-                    text: 'PREVIEW',
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                TextSpan(
-                    text: ' mode to test the link (optional). Then tap on '),
-                TextSpan(
-                    text: 'Share profile',
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                TextSpan(text: ' to get your profile link.\n'),
-              ],
-            ),
-          ),
+              'Once you\'re done. Switch to **PREVIEW** mode to test the link (optional). Then tap on **Share profile** to get your profile link.'),
           Text('\nNote: Changes are saved and synced automatically.',
               style:
                   TextStyle(fontStyle: FontStyle.italic, color: Colors.teal)),
@@ -67,9 +26,7 @@ class HelpDialogs {
       actions: [
         TextButton(
           child: const Text('Got it!\n'),
-          onPressed: () {
-            Navigator.pop(context);
-          },
+          onPressed: () => Navigator.pop(context),
         )
       ],
     );
@@ -85,31 +42,8 @@ class HelpDialogs {
         mainAxisSize: MainAxisSize.min,
         children: const [
           UnorderedListItem(
-            TextSpan(
-              children: [
-                TextSpan(
-                    text: 'Tap card',
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                TextSpan(
-                    text:
-                        ' open link. If some of the link didn\'t work as expected, you can edit the link by switching back to '),
-                TextSpan(
-                    text: 'EDITING',
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                TextSpan(text: ' page.\n')
-              ],
-            ),
-          ),
-          UnorderedListItem(
-            TextSpan(
-              children: [
-                TextSpan(text: 'Try to play around with the'),
-                TextSpan(
-                    text: ' Dough effect',
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-              ],
-            ),
-          ),
+              '**Tap card** open link. If some of the link didn\'t work as expected, you can edit the link by switching back to **EDITING** page.\n'),
+          UnorderedListItem('Try to play around with the **Dough effect**'),
         ],
       ),
       contentPadding: const EdgeInsets.fromLTRB(24.0, 20.0, 24.0, 2.0),
@@ -126,19 +60,22 @@ class HelpDialogs {
 }
 
 class UnorderedListItem extends StatelessWidget {
-  const UnorderedListItem(this.text, {Key? key}) : super(key: key);
-  final TextSpan text;
+  const UnorderedListItem(this.data, {Key? key}) : super(key: key);
+  final String data;
 
   @override
   Widget build(BuildContext context) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        const Text("• "),
-        Expanded(
-          child: Text.rich(text),
-        ),
-      ],
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 4),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          const Text("• "),
+          Expanded(
+            child: MarkdownBody(data: data),
+          ),
+        ],
+      ),
     );
   }
 }
